surya: added custom libcameraservice.so
Due to a bug in the device camera, timestamp cannot be acquired during
video recording. this leads to very strange videos which are hours long
but only the last few seconds actually contains any video.

To fix this bug a custom libcameraservice had to be adopted.
This code modification is detailed in the device tree commit: 9c3858367b89de4c17f5ce7edd6ecca464b1f453